refactor: Make a new module for version-specific functions
